Moved test_ols to test_array.py
